requirements.txt
matplotlib 
scikit-learn 
numpy 
keras==2.2.2 
tensorflow
jupyter
Pillow
flask_restful
